+++ src/common/js/index.js
@@ -3,10 +3,9 @@
  */
 
 import './global.js';
-import './config.js';
-//
 import './constants.js';
-import './defaults.js';
+import './configure.js';
+//
 import './functions.js';
 import './i18n.js';
 import './startup.js';